unify handling of error reponses to a single function
